Refactor out a common <BalanceAmountText /> component
Based on the code found in <TransactionListItemAmount />
