+++ frontend/src/Views/TransactionList/TransactionListItemAmount.tsx
@@ -1,44 +1,15 @@
-import {
-  createStyles,
-  ListItemSecondaryAction,
-  Theme,
-  Typography,
-  WithStyles,
-  withStyles
-} from "@material-ui/core";
+import { ListItemSecondaryAction } from "@material-ui/core";
 import React from "react";
+import BalanceAmountText from "../../Components/BalanceAmountText";
 
-const styles = (theme: Theme) =>
-  createStyles({
-    amountText: {
-      fontWeight: 500,
-      margin: 16
-    },
-    expense: {
-      color: "#f6787f"
-    },
-    income: {
-      color: "#68bcbe"
-    }
-  });
-
-interface IProps extends WithStyles<typeof styles> {
+interface IProps {
   amount: number;
 }
 
-const TransactionListItemAmount = ({ amount: amount, classes }: IProps) => {
-  const spending = amount < 0;
-  const amountText =
-    (spending ? "-" : "+") + " " + Math.abs(amount).toFixed(2) + "€";
-  return (
-    <ListItemSecondaryAction>
-      <Typography className={classes.amountText}>
-        <span className={spending ? classes.expense : classes.income}>
-          {amountText}
-        </span>
-      </Typography>
-    </ListItemSecondaryAction>
-  );
-};
+const TransactionListItemAmount = ({ amount: amount }: IProps) => (
+  <ListItemSecondaryAction>
+    <BalanceAmountText amount={amount} difference={true} />
+  </ListItemSecondaryAction>
+);
 
-export default withStyles(styles)(TransactionListItemAmount);
+export default TransactionListItemAmount;
